+++ src/lcl-str.h
@@ -14,7 +14,7 @@ void lcl_str_free(lcl_str *str);
 int lcl_str_reserve(lcl_str *st, size_t need);
 int lcl_str_putc(lcl_str *st, int c);
 int lcl_str_puts(lcl_str *st, const char *s);
-int lcl_str_write(lcl_str *buf, const char *str, int n);
+int lcl_str_write(lcl_str *buf, const char *str, size_t n);
 void lcl_str_append(lcl_str *str, const char *s, size_t n);
 
 #endif
